+++ app/Http/Controllers/Setup/Feature/FeatureController.php
@@ -17,6 +17,7 @@ use Illuminate\Support\Facades\Input;
 use InterventionImage;
 use App\Core\FormatGenerator As FormatGenerator;
 use App\Core\ReturnMessage As ReturnMessage;
+use App\Core\Check;
 
 class FeatureController extends Controller
 {
@@ -152,9 +153,26 @@ class FeatureController extends Controller
     public function destroy(){
         $id         = Input::get('selected_checkboxes');
         $new_string = explode(',', $id);
+        $delete_flag = true;
         foreach($new_string as $id){
-            $this->repo->delete($id);
+          $check = Check::checkToDelete("h_feature","feature_id",$id);
+
+          if(isset($check) && count($check)>0){
+              alert()->warning('This feature is used in hotel setup and you cannot delete it!')->persistent('OK');
+              $delete_flag = false;
+          }
+          else{
+              $this->repo->delete($id);
+          }
         }
-        return redirect()->action('Setup\Feature\FeatureController@index'); //to redirect listing page
+        if($delete_flag){
+            return redirect()->action('Setup\Feature\FeatureController@index')
+                ->withMessage(FormatGenerator::message('Success', 'Feature is deleted ...'));
+        }
+        else{
+            return redirect()->action('Setup\Feature\FeatureController@index')
+                ->withMessage(FormatGenerator::message('Fail', 'Feature is not deleted ...'));
+        }
+        // return redirect()->action('Setup\Feature\FeatureController@index'); //to redirect listing page
     }
 }
