+++ app/Http/Controllers/Setup/HotelNearby/HotelNearbyController.php
@@ -16,6 +16,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Input;
 use Auth;
+use App\Core\Check;
 
 class HotelNearbyController extends Controller
 {
@@ -111,9 +112,27 @@ class HotelNearbyController extends Controller
     public function destroy(){
         $id         = Input::get('selected_checkboxes');
         $new_string = explode(',', $id);
+        $delete_flag = true;
         foreach($new_string as $id){
-            $this->repo->delete($id);
+          // $check = $this->repo->checkToDelete($id);
+          $check = Check::checkToDelete("h_nearby","nearby_id",$id);
+
+          if(isset($check) && count($check)>0){
+              alert()->warning('This nearby place is used in hotel setup and you cannot delete it!')->persistent('OK');
+              $delete_flag = false;
+          }
+          else{
+              $this->repo->delete($id);
+          }
         }
-        return redirect()->action('Setup\HotelNearby\HotelNearbyController@index'); //to redirect listing page
+        if($delete_flag){
+            return redirect()->action('Setup\HotelNearby\HotelNearbyController@index')
+                ->withMessage(FormatGenerator::message('Success', 'Hotel Nearby is deleted ...'));
+        }
+        else{
+            return redirect()->action('Setup\HotelNearby\HotelNearbyController@index')
+                ->withMessage(FormatGenerator::message('Fail', 'Hotel Nearby is not deleted ...'));
+        }
+        // return redirect()->action('Setup\HotelNearby\HotelNearbyController@index'); //to redirect listing page
     }
 }
