+++ app/Http/Controllers/Setup/Landmark/LandmarkController.php
@@ -15,6 +15,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use Auth;
 use Illuminate\Support\Facades\Input;
+use App\Core\Check;
 
 class LandmarkController extends Controller
 {
@@ -120,9 +121,26 @@ class LandmarkController extends Controller
     public function destroy(){
         $id         = Input::get('selected_checkboxes');
         $new_string = explode(',', $id);
+        $delete_flag = true;
         foreach($new_string as $id){
-            $this->repo->delete($id);
+          $check = Check::checkToDelete("h_landmark","landmark_id",$id);
+
+          if(isset($check) && count($check)>0){
+              alert()->warning('This landmark is used in hotel setup and you cannot delete it!')->persistent('OK');
+              $delete_flag = false;
+          }
+          else{
+              $this->repo->delete($id);
+          }
         }
-        return redirect()->action('Setup\Landmark\LandmarkController@index'); //to redirect listing page
+        if($delete_flag){
+            return redirect()->action('Setup\Landmark\LandmarkController@index')
+                ->withMessage(FormatGenerator::message('Success', 'Landmark is deleted ...'));
+        }
+        else{
+            return redirect()->action('Setup\Landmark\LandmarkController@index')
+                ->withMessage(FormatGenerator::message('Fail', 'Landmark is not deleted ...'));
+        }
+        // return redirect()->action('Setup\Landmark\LandmarkController@index'); //to redirect listing page
     }
 }
