+++ app/Http/Controllers/Setup/Township/TownshipController.php
@@ -18,6 +18,7 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Input;
 use App\Core\FormatGenerator As FormatGenerator;
 use App\Core\ReturnMessage As ReturnMessage;
+use App\Core\Check;
 
 class TownshipController extends Controller
 {
@@ -119,10 +120,28 @@ class TownshipController extends Controller
     public function destroy(){
         $id         = Input::get('selected_checkboxes');
         $new_string = explode(',', $id);
+        $delete_flag = true;
         foreach($new_string as $id){
-            $this->repo->delete($id);
+          //check in hotel setup whether this township is used or not
+          $hotel_setup_check = Check::checkToDelete("hotels","township_id",$id);
+
+          if(isset($hotel_setup_check) && count($hotel_setup_check)>0){
+            alert()->warning('This township is used in hotel setup and you cannot delete it!')->persistent('OK');
+            $delete_flag = false;
+          }
+          else{
+              $this->repo->delete($id);
+          }
         }
-        return redirect()->action('Setup\Township\TownshipController@index'); //to redirect listing page
+        if($delete_flag){
+            return redirect()->action('Setup\Township\TownshipController@index')
+                ->withMessage(FormatGenerator::message('Success', 'Township is deleted ...'));
+        }
+        else{
+            return redirect()->action('Setup\Township\TownshipController@index')
+                ->withMessage(FormatGenerator::message('Fail', 'Township is not deleted ...'));
+        }
+        // return redirect()->action('Setup\Township\TownshipController@index'); //to redirect listing page
     }
